New option "-o" to directly generate assembly object
By default, file data contents are generated by initializing a static array.
However, data can be generated directly in ivm64 assembly object format which
does not need to be compiled (only linked).
To this end, invoke this script with "-o out.o" as first two arguments:

    ivm64-fsgen -o out.o $(find folder_name) # assembly object 'out.o' is generated
                                             # with the filesystem in it

In case of very large filesystem (>4GB), it is recommended to generate data in
assembly object format, and link it directly with the rest of the project:
    # Directly, generate the assembly object with the files of interest:
       ivm64-fsgen -o ivmfs-large.o $(find folder_name_with_very_large_files)
    # Optionally, generate an empty filesystem (in C, contaning fs primitives):
       ivm64-fsgen > ivmfs-empty.c
    # Link ivmfs-large.o  with the rest of the project (optional with an empty fs)
    # Do not forget to place the object with the FS in the first place
       ivm64-gcc ivmfs-large.o ivmfs-empty.c main.c ....